frontend/src/aux/: rename institutionConstants.js -> institutionInfo.js
