+++ frontend/src/aux/helpers.test.js
@@ -1,5 +1,5 @@
 import { getParameterFromQueryString } from './helpers';
-import { bannerInstitutionInfo } from './institutionConstants';
+import { bannerInstitutionInfo } from './institutionInfo';
 
 const institutionNamesUpperCase = Object.keys(bannerInstitutionInfo).map(institutionName => institutionName.toUpperCase());
 describe.each(institutionNamesUpperCase)(
