+++ frontend/src/components/Banner/Banner.js
@@ -2,7 +2,7 @@ import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import { getInstitutionQueryParameter } from '../../aux/helpers';
-import { bannerInstitutionInfo } from '../../aux/institutionConstants';
+import { bannerInstitutionInfo } from '../../aux/institutionInfo';
 
 const Banner = () => {
   const institutionName = getInstitutionQueryParameter();
